Add Outgoing Pallets Loading routes to web.php
- Introduced new routes for managing outgoing pallets loading, including vehicle details, allocations, pallet selection, and AI planning.
- Added methods in OutgoingPalletsLoadingController to handle the new functionalities.

+++ martini/database/migrations/2026_02_19_141754_create_outgoing_pallet_types.php
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->timestamps();
         });
         Schema::connection('tandc_live')->table('outgoing_pallet', function (Blueprint $table) {
-            $table->unsignedBigInteger('outgoing_pallet_type_id')->default(1)->after('id');
+            $table->integer('outgoing_pallet_type_id')->default(1)->after('id');
             $table->date('estimated_delivery_date')->after('outgoing_pallet_type_id');
             $table->boolean('dispatched')->default(false)->after('estimated_delivery_date');
         });
